Refactor into a seperate LocationSelect component
