+++ clients/sample-app/docker-compose.yml
@@ -20,7 +20,7 @@ services:
     container_name: fides-postgres-example-db
     image: postgres:12
     healthcheck:
-      test: [ "CMD-SHELL", "pg_isready -U postgres" ]
+      test: ["CMD-SHELL", "pg_isready -U postgres"]
       interval: 15s
       timeout: 5s
       retries: 5
