+++ clients/sample-app/src/components/Header/style.module.scss
@@ -11,20 +11,3 @@
     margin: 12px;
   }
 }
-
-.select {
-  position: absolute;
-  display: flex;
-  right: 0;
-  margin: 2em;
-  min-width: 200px;
-}
-
-.option {
-  display: flex;
-
-  img {
-    margin: 0;
-    margin-right: 0.5em;
-  }
-}
