+++ clients/sample-app/src/components/LocationSelect/style.module.scss
@@ -1,0 +1,16 @@
+.select {
+  position: absolute;
+  display: flex;
+  right: 0;
+  margin: 2em;
+  min-width: 200px;
+}
+
+.option {
+  display: flex;
+
+  img {
+    margin: 0;
+    margin-right: 0.5em;
+  }
+}
