+++ clients/sample-app/src/pages/index.tsx
@@ -24,21 +24,21 @@ const IndexPage = ({ products }: Props) => {
     fidesScriptTagUrl += `?location=${location}`;
   }
   return (
-  <>
-    <Head>
-      <title>Cookie House</title>
-      <meta
-        name="description"
-        content="Sample Project used within Fides (github.com/ethyca/fides)"
-      />
-      <link rel="icon" href="/favicon.ico" />
-      <link rel="stylesheet" href="https://rsms.me/inter/inter.css" />
-      <script src={fidesScriptTagUrl} />
-    </Head>
+    <>
+      <Head>
+        <title>Cookie House</title>
+        <meta
+          name="description"
+          content="Sample Project used within Fides (github.com/ethyca/fides)"
+        />
+        <link rel="icon" href="/favicon.ico" />
+        <link rel="stylesheet" href="https://rsms.me/inter/inter.css" />
+        <script src={fidesScriptTagUrl} />
+      </Head>
 
-    <Home products={products} />
-  </>
+      <Home products={products} />
+    </>
   );
-}
+};
 
 export default IndexPage;
